REFACTOR useMediaQuery custom hook
